Fix issue #10 -- genome bloat, not a display error
Both issues were related. I had changed the way
we handle the spawn pipeline, but had forgotten
to remove an obsolete mechanism, (d'oh!), which
was storing up genomes and never releasing them.

It turns out that the debug display was correct:
we were carrying around a bunch of stale genomes.